requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url requirements/base.in
#
attrs==20.3.0
    # via jsonschema
certifi==2020.12.5
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.14.6
    # via cryptography
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==3.4.8
    # via
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
django-admin-index==1.5.0
    # via -r requirements/base.in
django-axes==5.14.0
    # via -r requirements/base.in
django-better-admin-arrayfield==1.4.2
    # via mozilla-django-oidc-db
django-choices==1.7.1
    # via
    #   -r requirements/base.in
    #   vng-api-common
django-compat==1.0.15
    # via django-hijack-admin
django-filter==2.4.0
    # via
    #   -r requirements/base.in
    #   vng-api-common
django-formtools==2.3
    # via maykin-django-two-factor-auth
django-hijack-admin==2.1.10
    # via -r requirements/base.in
django-hijack==2.3.0
    # via
    #   -r requirements/base.in
    #   django-hijack-admin
django-ipware==3.0.2
    # via django-axes
django-jsonsuit==0.4.4
    # via -r requirements/base.in
django-ordered-model==3.4.1
    # via django-admin-index
django-otp==1.0.6
    # via maykin-django-two-factor-auth
django-phonenumber-field==5.2.0
    # via maykin-django-two-factor-auth
django-redis==4.12.1
    # via -r requirements/base.in
django-rest-framework-condition==0.1.1
    # via vng-api-common
django-rosetta==0.9.6
    # via -r requirements/base.in
django-sniplates==0.7.0
    # via -r requirements/base.in
django-solo==1.1.5
    # via
    #   mozilla-django-oidc-db
    #   sharing-configs
    #   vng-api-common
django==2.2.28
    # via
    #   -r requirements/base.in
    #   django-axes
    #   django-choices
    #   django-filter
    #   django-formtools
    #   django-hijack
    #   django-jsonsuit
    #   django-otp
    #   django-phonenumber-field
    #   django-redis
    #   django-rest-framework-condition
    #   django-rosetta
    #   django-sniplates
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   drf-yasg
    #   maykin-django-two-factor-auth
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   sharing-configs
    #   vng-api-common
djangorestframework-camel-case==1.2.0
    # via vng-api-common
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   drf-nested-routers
    #   drf-spectacular
    #   drf-yasg
    #   vng-api-common
drf-nested-routers==0.93.3
    # via
    #   -r requirements/base.in
    #   vng-api-common
drf-spectacular==0.17.2
    # via -r requirements/base.in
drf-yasg==1.20.0
    # via vng-api-common
elastic-apm==6.1.1
    # via -r requirements/base.in
gemma-zds-client==1.0.0
    # via vng-api-common
idna==2.10
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-yasg
iso-639==0.4.5
    # via vng-api-common
isodate==0.6.0
    # via vng-api-common
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
josepy==1.9.0
    # via mozilla-django-oidc
jsonschema==3.2.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
markupsafe==1.1.1
    # via jinja2
maykin-django-two-factor-auth[phonenumbers]==2.0.3
    # via -r requirements/base.in
mozilla-django-oidc-db==0.7.2
    # via -r requirements/base.in
mozilla-django-oidc==1.2.4
    # via mozilla-django-oidc-db
oyaml==1.0
    # via vng-api-common
packaging==20.9
    # via drf-yasg
phonenumbers==8.12.31
    # via maykin-django-two-factor-auth
pillow==9.0.1
    # via -r requirements/base.in
polib==1.1.1
    # via django-rosetta
psycopg2==2.8.6
    # via
    #   -r requirements/base.in
    #   mozilla-django-oidc-db
pycparser==2.20
    # via cffi
pyjwt==2.0.1
    # via
    #   gemma-zds-client
    #   vng-api-common
pyopenssl==20.0.1
    # via josepy
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
python-decouple==3.4
    # via -r requirements/base.in
python-dotenv==0.17.0
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   django
pyyaml==5.4.1
    # via
    #   drf-spectacular
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
qrcode==6.1
    # via maykin-django-two-factor-auth
redis==3.5.3
    # via django-redis
requests==2.25.1
    # via
    #   coreapi
    #   django-rosetta
    #   gemma-zds-client
    #   mozilla-django-oidc
    #   sharing-configs
    #   vng-api-common
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.4
    # via drf-yasg
sentry-sdk==1.0.0
    # via -r requirements/base.in
sharing-configs==0.1.1
    # via -r requirements/base.in
six==1.15.0
    # via
    #   django-choices
    #   django-compat
    #   django-rosetta
    #   isodate
    #   jsonschema
    #   mozilla-django-oidc
    #   pyopenssl
    #   qrcode
sqlparse==0.4.2
    # via django
unidecode==1.2.0
    # via vng-api-common
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-spectacular
    #   drf-yasg
urllib3==1.26.6
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements/base.in
vng-api-common==1.6.2
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
